+++ src/lib/printk.c
@@ -52,10 +52,10 @@ size_t vsprintk(char *buf, const char *fmt, va_list args)
             switch (*fmt) {
                 case 'x': {
                     unsigned long number = is_long ? va_arg(args, unsigned long) : va_arg(args, unsigned int);
-                    int length = is_long ? 16 : 8;
-                    int length_in_bits = is_long ? 64 : 32;
-                    int byte = 0;
-                    int i = 0;
+                    size_t length = is_long ? 16 : 8;
+                    size_t length_in_bits = is_long ? 64 : 32;
+                    uint8_t byte = 0;
+                    size_t i = 0;
                     bool keep_zeros = false;
 
                     for (i = 0; i < length; i++) {
